requirements.txt
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.8.1
certifi==2025.4.26
Django==4.2.22
django-ninja==1.4.3
exceptiongroup==1.3.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
iniconfig==2.1.0
packaging==25.0
pluggy==1.6.0
psycopg==3.2.9
psycopg-binary==3.2.9
pydantic==2.11.5
pydantic_core==2.33.2
Pygments==2.19.1
PyJWT==2.10.1
pytest==8.4.0
pytest-django==4.11.1
python-telegram-bot==22.1
sniffio==1.3.1
sqlparse==0.5.3
tomli==2.2.1
typing-inspection==0.4.1
typing_extensions==4.14.0
httpx==0.28.1
